+++ CHANGES.md
@@ -116,6 +116,10 @@ Changes in 10.0
 - CSS WPadmin fix menu select width in stylesheet.css
 - Enrollment Start: No N/A option for new student in StudentUsersInfo.fnc.php
 
+Changes in 9.3.2
+----------------
+- Fix regression since 9.2.1 fields other type than Select Multiple from Options in CategoryBreakdownTime.php
+
 Changes in 9.3.1
 ----------------
 - Fix regression since 2.9 Schedule multiple courses in plugins/Moodle/Scheduling/MassSchedule.php